Joystick: platform independent interface
This is essentially the same API that SDL2 and GLFW support. It does not have
the limitation of a max number of joysticks. It's possible that we may want to
support a max number of joysticks, because that can avoid an allocation, but
there's not really a significant amount of allocation that can happen due to
this!

I've trawled docs for all the major platforms and this API should be fine.
SDL2 emulates both a poll and event-based interface, by either polling
joystick state during the event loop, or updating a local copy of joystick
state and returning that for the poll.