impl set_cursor_state
grab mode use pointerlock https://www.w3.org/TR/pointerlock/
this behavior is not the same as other backend.

I wanted this because as set_cursor_position is not available on
emscripten backend. the only way to have a FPS camera is to use
pointerlock. So you can use grab to have FPS camera on emscripten.

also show_mouse is implement in javascript as in a pull request that
hasn't been merged https://github.com/kripken/emscripten/pull/4616