Take `WindowBuilder` rather than `Window` to `ContextBuilder::build`
Although this is slightly less flexible than taking the `Window`
directly for the user, this does solve the X11 context creation order
problem and also solves the problem where a user might unsafely create a
GL context for a window already associated with a directX context.

This only updates the API and does not yet update platform-specific
code. See the following commits.