Fix context creation on windows
The windows backend creates a new dummy window for loading context specific functions.
This dummy window reuses the existing class of the real window, but this interferes with thread-local storage, which are not initialized on the main thread.
Registering a new dummy specific class with the default callback for event handling resolved this issue.
See #906 for more details