Add Apple Silicon support
Apple Silicon support was added to winit in the latest (unreleased)
master version. It also renamed some types, so we'll need to do the
same here to keep compatible.

The dependency update won't be the `git` version ultimately, of course,
but I'm adding it in for now in case anyone else wants to know how
to build on M1 in the meantime.