Fix context sharing EGL/GLX mismatch
The code which determines whether to try EGL or GLX first does not
consider whether an existing context has been passed for context
sharing.

Since the logic for determining whether to try EGL or GLX first depends
on various factors (such as whether it's a headless context or not),
which one it tries first can differ from one invocation to the next.

On systems where both EGL and GLX are available, this can cause a panic
when the code which tries to create an EGL or GLX context suddenly
notices that a context already exists, and it is not of the requested
type.

Fix this by ignoring the `prefer_egl` and `force_prefer_unless_only`
parameters entirely if we have a shared context, since there is only one
option which can work.

The new `expect(...)` calls and the existing `panic!(...)` calls which
have been upgraded with a descriptive message are now guaranteed to
never happen as far as I can tell. They are purely to guard against
future code changes and to serve as documentation in case anybody
wonders why EGL is not being checked like in the cases further down.