Switch from `ndk-glue` to `raw-window-handle` crate
`winit` provides a `raw-window-handle` specifically for crates to
process window pointers/references in a crate-agnostic way.  `glutin`
can use it too, to get rid of its `ndk-glue` dependency which otherwise
has a hard versioning dependency  as it should be updated in sync with
`winit`.

Over time this should be worked into the other backends, in turn
removing `winit` from the `new_windowed()` function signature and
replacing it with `impl HasRawWindowHandle` or `dyn HasRawWindowHandle`.
This will also allow Android users to pass generic `Surface`s (first
converted to a so-called `NativeWindow` in the NDK through [#272], then
passed as `RawWindowHandle` thanks to [#274]) created in Java / Kotlin /
Flutter to render to individual UI elements instead of the entire screen
with a `NativeActivity` (`ndk-glue`).  It'll in turn make the code more
generic, too, as `raw-window-handle` has a variant for every platform /
windowing system currently supported by glutin.

[#272]: https://github.com/rust-windowing/android-ndk-rs/pull/272
[#274]: https://github.com/rust-windowing/android-ndk-rs/pull/274