Replace hardcoded `html` links with intradoc-links
Intradoc links are more concise to write and checked by `rustdoc` for
validity.  It'll not be allowed to create a link to an item that doesn't
exist - something rustdoc can't really check with (relative or not)
links to "random" html pages.

For example, this change helped find dead links to `RawContextExt` whose
module was renamed from `os` to `platform_impl` + `platform` well over 3
years ago, yet the links weren't updated.

At the same time intradoc-links receive a nice highlight from
rust-analyzer, and can be navigated to using jump-to-symbol.