api/egl: Add EGLDevice wrappers
This allows querying the available EGL devices on the system and getting the
device used by the display. A device can provide information such as the name
of the device the vendor and in the future, other info such as the DRM device
node on Linux. Mesa does not implement the extensions at the moment to allow
getting the name and vendor of the device.

A headless Display can be created using a Device. Along with this addition, an
EGL specific api to make a context current without a surface is also available.