glx: rework error handling
With GLX the error handling is done by the drivers by getting the
error hook from the Xlib and manually invoking. So this PR changes
handling to account for that, by wrapping glx calls which could
result in errors inside the RAII callbacks.

See https://github.com/emilk/egui/issues/2661 for some details.