egl/surface: Don't dereference native window pointer before passing to legacy `CreateWindowSurface`
Just like for KHR and EXT the native window _is_ `egl::NativeWindowType`
(which is a `*const c_void` under the hood).  By casting it to _a
pointer of_ `egl::NativeWindowType` instead, and derferencing that, we'd
be reading the first few bytes of whatever this opaque handle was
pointing to (if it even is a valid pointer) and segfaulting on at least
Android inside `libEGL`.  In other words, never dereference opaque
handles which may not even be a valid memory address in the first place.