egl/surface: fix window deref with the legacy API
Make handling of `NativeWindow` and `NativePixmap` type safe by
removing `as _` casts when passing to EGL functions accepting pointers.

The `eglNativeWindow` type is a pointer on all the platforms and
requires passing as is, except X11 where with the legacy functions it's
XID(ulong) and with KHR and EXT extensions you must pass the `*mut XID`
since those require an actual pointer.

This bug was introduced by the commit 8bd0eadba43a (egl: fix call to
`eglCreate{Window,Pixmap}Surface), where it fixed a crash on X11 and
added another crash on both Android and Win32.

Co-authored-by: Kirill Chibisov <contact@kchibisov.com>
Fixes: #1584.