chore: remove `.*SurfaceAccessor` traits
These traits were hacks to avoid the lack of GATs and provide a
detached trait with generic. Given that glutin MSRV is now `1.65`
which has GATs, we could safely remove said traits. Their methods
simply got moved to `.*GlContext` traits, where they originally
belonged to.

While the change is breaking, normal user code which was using
`prelude` is not affected by it.