egl/display: Pass *pointer to* output variable when querying `DEVICE_EXT`
By casting a `null_mut()` ptr and passing it directly _by value_ to
`QueryDisplayAttribEXT()`, the function ignores writing the output as
it didn't receive an address (it received `NULL`) to write the device
to.  Instead we should take the _pointer address of this `null_mut()`
pointer_ and provide that to the function instead so that it can
_overwrite_ it with a pointer to the requested `eglDevice`.

This is even more clear when allocating a `MaybeUninit` in which the
pointer will be returned, which has a convenient `.as_mut_ptr()` member
to pass its pointer value directly into `QueryDisplayAttribEXT()`.